enum component
